requirements.txt
deepgram-sdk>=3.0.0
pyaudio
python-dotenv
websockets
